--- lib/mailing.php
@@ -1,4 +1,0 @@
-<?php
-    // Meteré aquí todo lo que tenga que ver con los mails para separarlo del tema usuarios.
-
-?>
