+++ lib/verifyResetPass.php
@@ -1,0 +1,11 @@
+<?php
+require_once 'users.php';
+
+if (isset($_GET)) {
+    $usuari['mail']=$_GET['mail'];
+    $usuari['activateCode']=$_GET['code'];
+    searchAccount($usuari);
+    
+}
+header("Location: ../index.php");
+?>
